+++ src/Helpers/StringHelpers.h
@@ -5,36 +5,42 @@
 #ifndef POMS_STRINGSPLITHELPER_H
 #define POMS_STRINGSPLITHELPER_H
 
-#include <string>
 #include <sstream>
+#include <string>
 #include <vector>
 
-template<class Elem>
-using tstring = std::basic_string<Elem, std::char_traits<Elem>, std::allocator<Elem>>;
+template <class Elem>
+using tstring =
+    std::basic_string<Elem, std::char_traits<Elem>, std::allocator<Elem>>;
 
-template<class Elem>
-using tstringstream = std::basic_stringstream<Elem, std::char_traits<Elem>, std::allocator<Elem>>;
+template <class Elem>
+using tstringstream =
+    std::basic_stringstream<Elem, std::char_traits<Elem>, std::allocator<Elem>>;
 
 class StringHelpers {
- public:
-  template<typename Elem>
-  static std::vector<tstring<Elem>> split(tstring<Elem> text, Elem const delimiter) {
+public:
+  template <typename Elem>
+  static std::vector<tstring<Elem>> split(tstring<Elem> text,
+                                          Elem const delimiter) {
     auto sstr = tstringstream<Elem>{text};
     auto tokens = std::vector<tstring<Elem>>{};
     auto token = tstring<Elem>{};
     while (std::getline(sstr, token, delimiter)) {
-      if (!token.empty()) tokens.push_back(token);
+      if (!token.empty())
+        tokens.push_back(token);
     }
 
     return tokens;
   }
 
-  template<typename Elem>
-  static std::vector<tstring<Elem>> split(tstring<Elem> text, tstring<Elem> const &delimiters) {
+  template <typename Elem>
+  static std::vector<tstring<Elem>> split(tstring<Elem> text,
+                                          tstring<Elem> const &delimiters) {
     auto tokens = std::vector<tstring<Elem>>{};
 
     size_t pos, prev_pos = 0;
-    while ((pos = text.find_first_of(delimiters, prev_pos))!=std::string::npos) {
+    while ((pos = text.find_first_of(delimiters, prev_pos)) !=
+           std::string::npos) {
       if (pos > prev_pos)
         tokens.push_back(text.substr(prev_pos, pos - prev_pos));
       prev_pos = pos + 1;
@@ -45,6 +51,20 @@ class StringHelpers {
 
     return tokens;
   }
+
+  // Helper function to concatenate elements of a string vector with a given
+  // delimiter
+  static std::string join(const std::vector<std::string> &vec,
+                          const std::string &delimiter) {
+    std::string result;
+    for (size_t i = 0; i < vec.size(); i++) {
+      result += vec[i];
+      if (i != vec.size() - 1) {
+        result += delimiter;
+      }
+    }
+    return result;
+  }
 };
 
-#endif //POMS_STRINGSPLITHELPER_H
+#endif // POMS_STRINGSPLITHELPER_H